requirements.txt
flask==3.0.3
pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
tqdm==4.66.5
google-cloud-storage==2.18.2
openpyxl==3.1.5
